+++ nessi/io/__init__.py
@@ -24,8 +24,7 @@ from __future__ import print_function
 from .su_fmt import SUdata
 from .su_fmt import sucreate
 from .su_fmt import susrcinv
-from .nessidata import DataStruct
-from .suformat import suread
+from .su_fmt import suread
 
 if __name__ == '__main__':
     import doctest