+++ nessi/io/suformat.py
@@ -30,14 +30,16 @@ def _check_endian(fname):
     nsb = np.frombuffer(btmp, dtype='>h', count=1, offset=114)[0]
     if(bsize%((nsl*4)+240) == 0):
         endian = 'l'
+        ntrac = int(bsize/((nsl*4)+240))
     else:
         if(bsize%((nsb*4)+240) == 0):
             endian = 'b'
+            ntrac = int(bsize/((nsb*4)+240))
         else:
             sys.exit("Unable to read "+fname+"\n")
     file.close()
 
-    return endian
+    return endian, ntrac
 
 def _sutype():
     """
@@ -78,6 +80,40 @@ def _sutype():
 
     return sudtype
 
+def _suhdr2dict():
+    """
+    Convert SU header in dictionnary for PANDAS DataFrame
+    """
+
+    # Dictionnary header
+    dheader = [{
+                # Identification
+                'id':np.nan,     # data identification code
+                # Timestamp
+                'date':np.nan,   # Date in YYYYMMDD format
+                'time':np.nan,   # Time in HHMMSS format
+                # Time domain data
+                'ns':np.nan,     # Number of time samples
+                'dt':np.nan,     # Time sampling
+                'delrt':np.nan,  # Delta time
+                # Data size if not time domain
+                'n1':np.nan,     # number of samples in the first dimension
+                'd1':np.nan,     # sampling in the first dimension
+                'f1':np.nan,     #
+                'n2':np.nan,     # number of sample in the second dimension
+                'd2':np.nan,     # sampling in the second dimension
+                'f2':np.nan,     #
+                # Acquisition
+                'sx':np.nan,     # source coordinate X
+                'sy':np.nan,     # source coordinate Y
+                'sz':np.nan,     # source coordinate Z (elevation)
+                'gx':np.nan,     # receiver coordinate X
+                'gy':np.nan,     # receiver coordinate Y
+                'gz':np.nan,     # receiver coordinate Z (elevation)
+                },]
+
+    return dheader
+
 def suread(fname):
     """
     Read Seismic Unix files and store in NESSI data structure.
@@ -89,7 +125,7 @@ def suread(fname):
     sudata = DataStruct()
 
     # Check endianess
-    endian = _check_endian(fname)
+    endian, ntrac = _check_endian(fname)
 
     # Open the file to read
     sufile = open(fname, 'rb')
@@ -102,17 +138,12 @@ def suread(fname):
         sudtype = _sutype()
         npdtype = '<f4'
 
-    # Get the header of the first trace (240 lenght)
+    # Get the header of the first trace (240 bytes)
     bhdr = sufile.read(240)
-
-    # Read the header of the first trace
     hdr = np.frombuffer(bhdr, dtype=sudtype, count=1)[0]
-
-    # Calculate the lenght of traces in bites
+    # Get the  the first trace data values
     btrc = sufile.read(hdr['ns']*4)
-
-    # Close SU file
-    sufile.close()
+    trc = np.frombuffer(btrc, dtype=(npdtype, hdr['ns']), count=1)[0]
 
     # Convert first trace to NESSI data structure
     if hdr['trid'] != 0:
@@ -145,5 +176,47 @@ def suread(fname):
     sudata.header['gx'] = hdr['gx']*fscalco
     sudata.header['gy'] = hdr['gy']*fscalco
     sudata.header['gz'] = hdr['gelev']*fscalel
+
+    if ntrac > 1:
+        for itrac in range(1, ntrac):
+            dheader = _suhdr2dict()
+            sudata.header = sudata.header.append(dheader, ignore_index=True)
+            # Get the header of the first trace (240 bytes)
+            bhdr = sufile.read(240)
+            hdr = np.frombuffer(bhdr, dtype=sudtype, count=1)[0]
+            # Get the  the first trace data values
+            btrc = sufile.read(hdr['ns']*4)
+            trc = np.frombuffer(btrc, dtype=(npdtype, hdr['ns']), count=1)[0]
+            # Convert first trace to NESSI data structure
+            if hdr['trid'] != 0:
+                sudata.header['id'][itrac] = hdr['trid']
+            else:
+                sudata.header['id'][itrac] = 10 # Seismic trace (default)
+
+            # Time domain data header values
+            sudata.header['ns'][itrac] = hdr['ns']
+            sudata.header['dt'][itrac] = hdr['dt']
+            sudata.header['delrt'][itrac] = hdr['delrt']
+            # Acquisition
+            scalco = hdr['scalco']
+            if scalco == 0:
+                fscalco = 1.
+            if scalco < 0:
+                fscalco = -1./np.float(scalco)
+            if scalco > 0:
+                fscalco = np.float(scalco)
+            scalel = hdr['scalel']
+            if scalel == 0:
+                fscalel = 1.
+            if scalel < 0:
+                fscalel = -1./np.float(scalel)
+            if scalel > 0:
+                fscalel = np.float(scalel)
+            sudata.header['sx'][itrac] = hdr['sx']*fscalco
+            sudata.header['sy'][itrac] = hdr['sy']*fscalco
+            sudata.header['sz'][itrac] = hdr['selev']*fscalel
+            sudata.header['gx'][itrac] = hdr['gx']*fscalco
+            sudata.header['gy'][itrac] = hdr['gy']*fscalco
+            sudata.header['gz'][itrac] = hdr['gelev']*fscalel
 
     return sudata